Move code that store log records in queue and work with env to another file
